generalize element rule head for sum encoding
